Change Carbon method to get last month date
If today is July 31th, this test fails, because the subMonth
method returns July 1st, instead of June 30th.

I dont really know how carbon works in this feature, so
the fastest solution is change how we get last month date
which is use today()->subDays(31), because this should
returns date of last month.